chore: move bridge status types to shared
